requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pdf2image==1.16.3
fpdf2==2.7.5
python-docx==0.8.11
comtypes==1.2.0
ffmpeg-python==0.2.0
gradio-client==0.2.9
apscheduler==3.10.1
requests==2.31.0